Create AccountBalCheck.sh
a brute-force password cracker for the MTN Cameroon mobile money website. It works by first getting the request headers from the website. Then, it iterates from 0 to 9999, generating a different password for each iteration. For each password, it makes a POST request to the login page, passing the username and password as parameters. If the request is successful, the script will get the OTP from the website and check if it is valid. If the OTP is valid, the script will get the account details and print them out